New dice updates, and camera fixes
Updated the scripts to be able to handle spawning of attackers and defenders dice. You can now retrieve the roll values of the attacker and defender individually, and specify an amount of dice to spawn and roll for both attacker and defender.

Also fixed the camera bug where the camera started with an orientation that was flat with the board plane. The camera will now start from a top-down view