Constructing the Gameloop
Pulled together all phases (Setup, Deployment, Attack, Fortify) into the gameloop. A lot of tinkering was necessary, however I did not dramatically change how each phase was designed. The most I changed in each was:
- variable names
- function parameters
- removing blocks of code that were obsolete
- adding my own bits of code to make sure everything works

I also implemented country highlighting in all phases of the game to indicate which countries are selectable at that current point in the game. This makes it MUCH easier to identify what to do, and where you can do it.

`Continent.cs` has been built on to track what players are in each continent, along with a function to return a bool stating whether a continent is fully owned by a player or not.

`Troop.cs` get built upon too, and now holds functions to retrieve the GameObject of each of the troop types and colour. You can also pass in a troop GameObject to a function and have it's corresponding troop colour returned.